Generate characteristics token and css

+++ packages/export-variables-rest-api/src/index.ts
@@ -1,6 +1,6 @@
 import * as https from "node:https";
 import {
-  readFile,
+  lstatSync,
   existsSync,
   mkdirSync,
   writeFileSync,
@@ -20,7 +20,14 @@ import {
   ThemeRoot,
   generateCssFromJson,
 } from "./tokenUtils.js";
-import { Color, GetVariableResponse, RGBA, Variable } from "./types.js";
+import {
+  Color,
+  GetFileNodesResponse,
+  GetPublishedStylesResponse,
+  GetVariableResponse,
+  RGBA,
+  Variable,
+} from "./types.js";
 import { stringifyRGBA } from "./colorUtils.js";
 
 // Make style dictionary aware of W3C file format, e.g. $value key
@@ -32,121 +39,176 @@ const figmaAccessToken = process.env.FIGMA_ACCESS_TOKEN;
 const TOKEN_PREFIX = "salt";
 const TOKEN_FOLDER = "./tokens";
 const CODE_OUTPUT_FOLDER = "./build";
-const LOCAL_EXAMPLE_FILE = "salt-example.json";
+const CHARACTERISTICS_TOKEN_FILE = "CHARACTERISTICS_TOKEN.json";
+const LOCAL_VARIABLE_EXAMPLE_FILE = "salt-example.json";
+const LOCAL_STYLE_NODES_EXAMPLE_FILE = "salt-style-nodes-example.json";
 // Leave blank if you don't want to save the response from REST API call. Give it a name so API response will be saved to be used in `loadLocalMockData`, e.g. LOCAL_EXAMPLE_FILE
-const SAVE_API_RESPONSE_PATH = LOCAL_EXAMPLE_FILE;
+const SAVE_VARIABLE_API_RESPONSE_PATH = LOCAL_VARIABLE_EXAMPLE_FILE;
+const SAVE_STYLE_NODES_API_RESPONSE_PATH = LOCAL_STYLE_NODES_EXAMPLE_FILE;
 
 const processData = processWithCustomCSSGeneration;
 
 /** Either call API or use local mock data */
-// callFigmaAPI(processData);
+// callFigmaVariableAPI(processData);
 loadLocalMockData(processData);
 
 function loadLocalMockData(
-  successCallback: (meta: GetVariableResponse["meta"]) => void
+  successCallback: (
+    meta: GetVariableResponse["meta"],
+    styleNodes: GetFileNodesResponse["nodes"]
+  ) => void
 ) {
-  readFile(LOCAL_EXAMPLE_FILE, { encoding: "utf8" }, (err, data) => {
-    if (err) {
-      console.error(err);
-      return;
-    }
-    const parsedData = JSON.parse(data);
-    successCallback(parsedData.meta);
+  const data = readFileSync(LOCAL_VARIABLE_EXAMPLE_FILE, { encoding: "utf8" });
+  const parsedData = JSON.parse(data);
+  const styleNodeData = readFileSync(LOCAL_STYLE_NODES_EXAMPLE_FILE, {
+    encoding: "utf8",
+  });
+  const parsedStyleNodeData = JSON.parse(styleNodeData);
+  successCallback(parsedData.meta, parsedStyleNodeData.nodes);
+}
+
+function callFigmaAPI(endpoint: string): Promise<string> {
+  return new Promise(function (resolve, reject) {
+    // The correct proxy `Agent` implementation to use will be determined
+    // via the `http_proxy` / `https_proxy` / `no_proxy` / etc. env vars
+    const agent = new ProxyAgent();
+
+    // HTTP request will now go through proxy
+    https.get(
+      `https://api.figma.com${endpoint}`,
+      {
+        agent,
+        headers: {
+          "X-FIGMA-TOKEN": figmaAccessToken,
+        },
+      },
+      (res) => {
+        // console.log(res.statusCode, res.headers);
+        // res.pipe(process.stdout);
+
+        const { statusCode } = res;
+        const contentType = res.headers["content-type"];
+
+        let error;
+        // Any 2xx status code signals a successful response but
+        // here we're only checking for 200.
+        if (statusCode !== 200) {
+          error = new Error(
+            "Request Failed to " +
+              endpoint +
+              ".\n" +
+              `Status Code: ${statusCode}`
+          );
+        } else if (!contentType || !/^application\/json/.test(contentType)) {
+          error = new Error(
+            "Invalid content-type.\n" +
+              `Expected application/json but received ${contentType}`
+          );
+        }
+        if (error) {
+          // Consume response data to free up memory
+          res.resume();
+          reject(error);
+          return;
+        }
+
+        res.setEncoding("utf8");
+        let rawData = "";
+        res.on("data", (chunk) => {
+          rawData += chunk;
+        });
+        res.on("end", () => {
+          try {
+            resolve(rawData);
+          } catch (e) {
+            reject(e);
+          }
+        });
+      }
+    );
   });
 }
 
-function callFigmaAPI(
-  successCallback: (meta: GetVariableResponse["meta"]) => void
+async function callFigmaVariableAPI(
+  successCallback: (
+    meta: GetVariableResponse["meta"],
+    styleNodes: GetFileNodesResponse["nodes"]
+  ) => void
 ) {
-  // The correct proxy `Agent` implementation to use will be determined
-  // via the `http_proxy` / `https_proxy` / `no_proxy` / etc. env vars
-  const agent = new ProxyAgent();
-
-  // HTTP request will now go through proxy
-  https.get(
-    `https://api.figma.com/v1/files/${figmaFileId}/variables/local`,
-    {
-      agent,
-      headers: {
-        "X-FIGMA-TOKEN": figmaAccessToken,
-      },
-    },
-    (res) => {
-      // console.log(res.statusCode, res.headers);
-      // res.pipe(process.stdout);
-
-      const { statusCode } = res;
-      const contentType = res.headers["content-type"];
-
-      let error;
-      // Any 2xx status code signals a successful response but
-      // here we're only checking for 200.
-      if (statusCode !== 200) {
-        error = new Error("Request Failed.\n" + `Status Code: ${statusCode}`);
-      } else if (!contentType || !/^application\/json/.test(contentType)) {
-        error = new Error(
-          "Invalid content-type.\n" +
-            `Expected application/json but received ${contentType}`
-        );
-      }
-      if (error) {
-        console.error(error.message);
-        // Consume response data to free up memory
-        res.resume();
-        return;
-      }
-
-      res.setEncoding("utf8");
-      let rawData = "";
-      res.on("data", (chunk) => {
-        rawData += chunk;
-      });
-      res.on("end", () => {
-        try {
-          if (SAVE_API_RESPONSE_PATH) {
-            writeFileSync(SAVE_API_RESPONSE_PATH, rawData);
-          }
-          const parsedData = JSON.parse(rawData);
-          successCallback(parsedData.meta);
-        } catch (e) {
-          console.error((e as Error).message);
-        }
-      });
-    }
+  const rawData = await callFigmaAPI(
+    `/v1/files/${figmaFileId}/variables/local`
   );
+
+  if (SAVE_VARIABLE_API_RESPONSE_PATH) {
+    writeFileSync(SAVE_VARIABLE_API_RESPONSE_PATH, rawData);
+  }
+  const { meta } = JSON.parse(rawData) as GetVariableResponse;
+
+  const styles = await getColorStylesFromAPI();
+  const styleNodes = await getNodeDetailFromAPI(styles.map((s) => s.node_id));
+
+  successCallback(meta, styleNodes);
 }
 
-function processWithCustomCSSGeneration(data: GetVariableResponse["meta"]) {
-  // console.debug("processWithCustomCSSGeneration", data);
+async function getColorStylesFromAPI() {
+  const rawData = await callFigmaAPI(`/v1/files/${figmaFileId}/styles`);
+  const parsedData = JSON.parse(rawData) as GetPublishedStylesResponse;
+  return parsedData.meta.styles;
+}
+
+async function getNodeDetailFromAPI(nodeIds: string[]) {
+  const rawData = await callFigmaAPI(
+    `/v1/files/${figmaFileId}/nodes?ids=${nodeIds.join(",")}`
+  );
+  if (SAVE_STYLE_NODES_API_RESPONSE_PATH) {
+    writeFileSync(SAVE_STYLE_NODES_API_RESPONSE_PATH, rawData);
+  }
+  const parsedData = JSON.parse(rawData) as GetFileNodesResponse;
+  return parsedData.nodes;
+}
+
+function processWithCustomCSSGeneration(
+  data: GetVariableResponse["meta"],
+  styleNodes: GetFileNodesResponse["nodes"]
+) {
+  console.debug("processWithCustomCSSGeneration", { data, styleNodes });
 
   // No need to add default to grouping using custom implementation
   const newData = updateApiResponse(data, { addDefault: false });
 
+  const charTokens = extractTokenFromStyles(newData, styleNodes);
+
   const tokens = extractTokenFromVariables(newData.variables);
+  // TODO: where to insert `palette` prefix? some name conflict e.g. accent
   // console.debug("extracted full tokens:", tokens);
 
-  writeTokensToFile(newData, tokens);
+  writeTokensToFile(newData, tokens, charTokens);
 
   generateCustomCSS();
 }
 
-function processWithStyleDictionary(data: GetVariableResponse["meta"]) {
+function processWithStyleDictionary(
+  data: GetVariableResponse["meta"],
+  styleNodes: GetFileNodesResponse["nodes"]
+) {
   console.log("processWithStyleDictionary", data);
 
   const newData = updateApiResponse(data, { addDefault: true });
   console.log("data with default name suffix", newData);
+  const charTokens = extractTokenFromStyles(newData, styleNodes);
 
   const tokens = extractTokenFromVariables(newData.variables);
   console.log("extracted full tokens:", tokens);
 
-  writeTokensToFile(newData, tokens);
+  writeTokensToFile(newData, tokens, charTokens);
 
   buildUsingStyleDictionary();
 }
 
 function writeTokensToFile(
   data: GetVariableResponse["meta"],
-  tokens: Record<string, Record<string, ThemeRoot>>
+  tokens: Record<string, Record<string, ThemeRoot>>,
+  charTokens: ThemeRoot
 ) {
   const allCollections = data.variableCollections;
 
@@ -174,6 +236,47 @@ function writeTokensToFile(
       console.log("Written token to", tokenFilePath);
     }
   }
+
+  const charTokenFilePath = join(TOKEN_FOLDER, CHARACTERISTICS_TOKEN_FILE);
+  writeFileSync(charTokenFilePath, JSON.stringify(charTokens, null, 2));
+
+  console.log("Written characteristics token to", charTokenFilePath);
+}
+
+function extractTokenFromStyles(
+  newData: GetVariableResponse["meta"],
+  styleNodes: GetFileNodesResponse["nodes"]
+) {
+  const fullTokens: ThemeRoot = {};
+
+  Object.values(styleNodes).forEach((node) => {
+    const filesBoundVars = node.document.boundVariables["fills"];
+    if (filesBoundVars && Array.isArray(filesBoundVars)) {
+      const boundVariable = newData.variables[filesBoundVars[0].id];
+      const tokenName = node.document.name;
+
+      let obj: any = fullTokens;
+
+      tokenName.split("/").forEach((groupName) => {
+        const camelGroupName = toCamelCase(groupName);
+        obj[camelGroupName] = obj[camelGroupName] || {};
+        obj = obj[camelGroupName] as any;
+      });
+
+      // TODO: resolvedType: "BOOLEAN" | "FLOAT" | "STRING" | "COLOR"
+      obj.$type = boundVariable.resolvedType === "COLOR" ? "color" : "number";
+      obj.$value = `{${boundVariable.name
+        .split("/")
+        .map(toCamelCase)
+        .join(".")}}`;
+    } else {
+      // No fills found?
+      throw new Error(
+        "No fills boundVariable found for node," + node.document.name
+      );
+    }
+  });
+  return fullTokens;
 }
 
 function extractTokenFromVariables(
@@ -250,28 +353,56 @@ function generateCustomCSS() {
    * - Otherwise add `.salt-theme`
    **/
 
-  for (const subfolderName of readdirSync(TOKEN_FOLDER)) {
-    const outputFolder = join(CODE_OUTPUT_FOLDER, subfolderName);
-    if (!existsSync(outputFolder)) {
-      mkdirSync(outputFolder, { recursive: true });
-    }
-    const tokenFolder = join(TOKEN_FOLDER, subfolderName);
-    for (const tokenFileName of readdirSync(tokenFolder)) {
-      // console.debug({ tokenFileName });
-      const tokenInput = readFileSync(join(tokenFolder, tokenFileName), {
+  for (const dirContent of readdirSync(TOKEN_FOLDER, { withFileTypes: true })) {
+    if (dirContent.isDirectory()) {
+      const outputFolder = join(CODE_OUTPUT_FOLDER, dirContent.name);
+      if (!existsSync(outputFolder)) {
+        mkdirSync(outputFolder, { recursive: true });
+      }
+      const tokenFolder = join(TOKEN_FOLDER, dirContent.name);
+      for (const tokenFileName of readdirSync(tokenFolder)) {
+        // console.debug({ tokenFileName });
+        const tokenInput = readFileSync(join(tokenFolder, tokenFileName), {
+          encoding: "utf-8",
+        });
+        const outputCss = generateCssFromJson(tokenInput, {
+          prefix: "salt",
+          rgbaFormat: true,
+          specialPrefixMap: SALT_SPECIAL_PREFIX_MAP,
+          removeSuffixDefault: true,
+          kebabCase: true,
+        });
+        // console.debug({ outputCss });
+        const outputFilePath = join(
+          outputFolder,
+          basename(tokenFileName, extname(tokenFileName)) + ".css"
+        );
+        // TODO: prettier
+        writeFileSync(outputFilePath, ":root {\n" + outputCss + "\n}\n", {
+          encoding: "utf8",
+        });
+        console.log("CSS output written to", outputFilePath);
+      }
+    } else if (
+      dirContent.isFile() &&
+      dirContent.name === CHARACTERISTICS_TOKEN_FILE
+    ) {
+      const tokenInput = readFileSync(join(TOKEN_FOLDER, dirContent.name), {
         encoding: "utf-8",
       });
+
       const outputCss = generateCssFromJson(tokenInput, {
         prefix: "salt",
         rgbaFormat: true,
         specialPrefixMap: SALT_SPECIAL_PREFIX_MAP,
         removeSuffixDefault: true,
         kebabCase: true,
+        ignoreNameSpecialPrefix: true,
       });
-      // console.debug({ outputCss });
+
       const outputFilePath = join(
-        outputFolder,
-        basename(tokenFileName, extname(tokenFileName)) + ".css"
+        CODE_OUTPUT_FOLDER,
+        basename(dirContent.name, extname(dirContent.name)) + ".css"
       );
       // TODO: prettier
       writeFileSync(outputFilePath, ":root {\n" + outputCss + "\n}\n", {
